RIP RPC and bitcoind integrations, move script.constants to bits.constants and script/ to script.py, start debugging test_bip143
