+++ conf/config.json
@@ -3,10 +3,6 @@
   "network": "mainnet",
   "datadir": "",
   "seeds": [],
-  "rpc_url": "",
-  "rpc_user": "",
-  "rpc_password": "",
-  "rpc_datadir": "",
   "input_format": "hex",
   "output_format": "hex"
 }
