+++ conf/config.toml
@@ -10,9 +10,3 @@ max_incoming_peers=25
 
 input_format="hex"
 output_format="hex"
-
-# connection to bitcoind daemon
-rpc_url=""
-rpc_user=""
-rpc_password=""
-rpc_datadir=""
