+++ src/bits/config.py
@@ -19,11 +19,6 @@ class Config(object):
         self.seeds = kwargs.get("seeds", [])
         self.max_outgoing_peers = kwargs.get("max_outgoing_peers", 5)
 
-        self.rpc_url = kwargs.get("rpc_url", "")
-        self.rpc_user = kwargs.get("rpc_user", "")
-        self.rpc_password = kwargs.get("rpc_password", "")
-        self.rpc_datadir = kwargs.get("rpc_datadir", "")
-
         self.input_format = kwargs.get("input_format", "hex")
         self.output_format = kwargs.get("output_format", "hex")
